feat: Домашнее задание 5.14. - Разделил файл на модули - Исправил ошибки и доработал код
