+++ js/util.js
@@ -1,0 +1,21 @@
+const createIdGenerator = () => {
+    let lastGeneratedId = 0;
+
+    return function () {
+        lastGeneratedId += 1;
+        return lastGeneratedId;
+    };
+};
+
+const getRandomInteger = (min, max) => {
+    const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+    const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+    const result = Math.random() * (upper - lower + 1) + lower;
+  
+    return Math.floor(result);
+};
+
+const getRandomArrayElement = (array) =>
+  array[getRandomInteger(0, array.length - 1)];
+
+export {getRandomInteger, getRandomArrayElement, createIdGenerator };
